Extract sync diagnose UI
